removing test_image_sscd and combining tests of sscd and pdq in test_image
